requirements.txt

flask
python-dotenv
pdfplumber
google-generativeai
